resolve global paths
